social-oauth-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# OAuth 2.0 client library
authlib==1.3.0
httpx==0.26.0
# OAuth 1.0a for Twitter media uploads
requests-oauthlib==1.3.1

# Token encryption
cryptography==42.0.0

# Background task scheduling
apscheduler==3.10.4
redis==5.0.1

# Social Media Platform SDKs
# Instagram/Facebook - Meta Graph API
requests==2.31.0
# Twitter/X - Tweepy
tweepy==4.14.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Image handling
Pillow==10.2.0
cloudinary==1.44.1

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Validation and security
email-validator==2.1.0
python-jose[cryptography]==3.3.0

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
